Split echobot and ChatHandler
